Add files via upload
I decided to play with a coffee order (simple) app idea. This currently has a menu array to pick a specific type of coffee, and a price for the coffee. I plan to add sizes later, and if you want to order more than one coffee you'll be able to.